+++ changelog.md
@@ -2,12 +2,18 @@
 
 # current
 
-Bug fix: year number must be specified when using day of year number as the time variable.
+- Added a general function to parse tabulated data. Other data reading functions will be deprecated.
+- Added date parsing options for `pandas.read_csv()`.
+- Bug fix: year number must be specified when using day of year number as the time variable.
 
 @TODO:
+- Improve the data reading functions.
+- Use `pandas.Timestamp` as the standard timestamp passed between functions.
+
 - Add timelag control in `chamber.yaml`
 - Test and refine the chamber schedule function for timelag optimization.
 - Daily summary plots.
+
 
 # 0.1.3 (2017-02-05)